requirements.txt
networkx>=3.0
numpy>=1.24.0
scipy>=1.10.0
sentence-transformers>=2.2.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dateutil>=2.8.0
icalendar>=5.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
flask>=3.0.0

# Optional: For interactive AI-enhanced input parsing
google-generativeai>=0.3.0
